Basic Button Functionality
Generating Different Problems. Input Answer and reveal. Need to work on checking if the answer is correct. A big issue I run into so far is multiple displays. Might have to create a for loop. Might want to work on CSS now so I can get a better idea for future tricks.