CSS Practice Design
Implemented Basic CSS on just the Practice div. It's sufficient. I need to do more work on basic website functions such as an indicator for a correct answer and whatnot. So far, for the examples part, I'm stuck on how to write those functions. I think I will have to do substr. But I only touched the tip of the parsing iceberg in C++, so I will have to learn more about it in Javascript.