Check answer. Visual improvements.  Reading individual character.
Check answer functional. However can't get it to check whenever there is user input. Nor can I keybind "enter" to the button. It seems to be a problem with the "addEventListener()".
Added some visuals along with the check answer.
Finally, I started some of the examples. I still have much to do. But first, I need to learn the syntax for arrays so I can store the individual character variables so I can use them later.