Replacement CSS
Only replaced Practice CSS design in 11.html. Will implement to the rest later. Used Flexbox and achieved the same look in a more efficient method. The end goal was to hopefully see some improvements on the mobile screen.  So far the mobile screen's main issue is the buttons. More specifically the text inside the buttons don't fit.