Universal classes. Starting Foiling Trick
Developed a basic universal class template for all of my tricks. I moved readChar() out of the classes to be shared. However there will need to be multiple readChar's based on the ammount of digits used inside the formula. ...Now that I think of it I could probobly just use a switch case to identify its type and then what their variables should be named as...

Also thinking about creating a carry function. It's definitely possible I just didn't use it in my 11 tricks class because carrys weren't prevelent and the parameters can often times be confusing. I will overcome it when developing the Foiling tricks since carrying is like a viral infection in these things.