CSS For dropdown box
Its functioning. BUT, I really dont like how huge the width of  "Problem Type" is. It makes the hover target so huge and its not needed. I might need to change it into an read-only input element so I can dynamically mess with its width without messing it up... But ill do that when I feel like touching css again.