Working on selected checkboxes displaying to display input
I have to work with arrays in order to randomize things. But the problem I'm facing with the arrays is that I only record the equation and not the entire problem. This is fine just for equation related things, but I need more information than just the equation in order for other functions to work. I can either get a function to evaluate it, or I can hopefully find a way to get the entire trick data...